+++ Changelog.md
@@ -4,6 +4,8 @@
 - Redirection `http` to `https` auto with use https vhost type
 - Generate self-signed certificates automatically
 - Clean create and permissions on logs directory
+- Add option `TORAN_HTTP_PORT` to allow arbitrary http port for nginx / toran proxy
+- Add option `TORAN_HTTPS_PORT` to allow arbitrary https port for nginx / toran proxy
 
 **1.5.1**
 - Upgrade toran proxy to version 1.5.1